pyproject.toml
[project]
name = "ai-music-composer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.7",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.5",
    "langchain-groq>=1.0.1",
    "langchain-ollama>=1.0.0",
    "music21>=9.9.1",
    "python-dotenv>=1.2.1",
    "scipy>=1.16.3",
    "streamlit>=1.51.0",
    "synthesizer>=0.2.0",
]
